Improved PropMap
I had originally had the Properties as params to Thing creation, but
that failed because the order of startup was ill-defined.  Then I'd made
then a val to override, but I'm realizing that that's very inconvenient
for loading a Space from the DB.

So made things subtler: they've back to being parameters, but are
evaluated lazily, so startup has a chance to work properly.