Finally got it right, I think
The heart of the problem has been ElemValue.  That has served as the
type "glue" between PType and Collection, and has been caused PType to
poison Collection with undesired type parmeters.

So I made the sad but necessary choice of making ElemValue truly
untyped.  This decouples everything else.  It means that an ElemValue is
only meaningful in the context of its PType, but that's probably correct
anyway, and it causes scads of type parameters to go away.