Fixed the relationship of Thing to Space
Realized, rather to my horror, that there was a fundamental flaw in how
Things relate to their Spaces.  I had a hacked and broken Thing.space
property; when I went to fix that, I realized that it can't *possibly*
work while maintaining the performance characteristics that I want,
because the Thing can't maintain a pointer to its Space if Spaces can
evolve efficiently.

So changed the whole attitude: Thing methods are, by and large,
evaluated using an implicit SpaceState.  You are required to say
deliberately which Space is the context.  This is a little more work,
but makes lots of problems go away.