You can now define Stylesheets as Things
I implemented CSS upload yesterday, but quickly realized that having to
re-upload for every edit is going to make me nuts.  So added an
alternative, of having a Stylesheet Model that you can link to, with a
CSS property containing the actual stylesheet text.  This is notably
much less efficient, but a heck of a lot easier to use.