Got Is a Model working again
Pulled at that string, and it came away in my hands.

Had to wrestle first with the fact that checkboxes -- the obvious way to
render a YesNo for editing -- don't send *anything* if they aren't
checked, which broke parsing.  So rewrote the page so that we can detect
properties that are defined but have no matching value returned.

Then, found that YesNo wasn't validating properly.  Found a bug there,
but then realized that validate() and doFromUser() are and should be
intimately related.  Made validate final, and doFromUser should now
throw input exceptions.

Then realized that parseBoolean is misleading and broken.  Wrote my own
version of it.