Lots of refactorings
Checkpoint: I'm in the middle of a huge refactor, on two principal
fronts.

First, I'm moving from URL-by-name as a special case to instead making
OIDs and Names largely interchangeable, which makes much more sense.

Second, I'm slowly introducing RequestContext, which encapsulates the
info we have on the incoming request, so I stop having to pass zillions
of parameters around.