Create and Edit now work better
Went to make them both ThingId-compatible.  That kind of came away in my
hands -- wound up refactoring a *lot* of code to make it possible.  But
things are moving in a good direction.

Introduced the ability to write directly from a ThingId, so that can now
become the key class in URL parts.

Fixed spaces.html to match the new model better.

Fixed spaceNav to prefer name-based URLs.
Same for things.html.

Added the ability to use space.anything on a ThingId, to centralize that
decision-making.