Progress towards being able to edit the Space
After putting the initial prototype up on the Web, discovered that
trying to edit the Space itself causes things to crash.  Gradually
working my way through a number of problems:

-- Displaying a flashed error caused a crash.  (Due to a miscast/match
problem.)
-- state.anything() didn't detect the Space itself by OID, which
resulted in returning SystemSpace instead.  (Ick!)
-- Similarly, anythingByName() didn't detect the Space by its name.
-- ModifyThing only worked with ordinary ThingStates, not other kinds.

Also, added code so that, for now, I log a stack trace on every flashed
error.  This is overkill, but useful in the short run.

Still have to add explicit code for other kinds, but getting towards the
right structure for that.