First pass of getting dynamic Properties working
Actually went much more easily than I expected -- still waiting for the
other shoe to drop.  But I've created a simple dynamic Property, used it
in a Thing, and it seems to be serializing/deserializing correctly.  UI
still sucks, of course, but that's much less scary.

Main complication was fixing the load-time dependency on SystemSpace.
Made Space load into a multi-phase affair, where only the Space itself
depends on SystemSpace, then the props depend on the Space, then the
Things depend on the props.  It's a bit rough, and will have problems
with meta-Properties, but it's a decent start.