Properties can now be Kind-specific
This helps clean up the Properties offered to you when creating or
editing -- each Property can declare that it only applies to Things or
Properties, and in that case it won't be offered as an option if you're
creating the wrong Kind.

Basically, meta-Properties are now properly declared as such, so they're
not in your face when creating Things.