Invented Modules
A day's worth of heavy refactoring.

The main point of the exercise was to create Modules, so that the code
stops spaghettizing quite so fast.

For the first Module, used Stylesheets, which are a nice bundle of
functionality.

As a consequence, wound up pulling PType input rendering into the Types
themselves.  I'm not sure this is correct, but it's much better-factored
than the way things were before.

Also introduced the notion of proper init and term for the first time.
That was inevitable.