Beginnings of QLText
Checkpointing, but have made significant progress.

Introduced the QLText type, and the beginnings of the QLParser.

Changed the TextTypes to be based on QLText (as they should be) instead
of Wikitext.

Introduced PlainTextType, for Properties that should have unvarnished
plain text.  This is a middle ground between Name and QLText.