We are now passing the entire RequestContext into rendering
We had been passing in the SpaceState, but I realized that this isn't
going to be enough in the long run.  Rendering gets into issues of
permission, so we also need to know the requester.  At which point, we
may as well just pass in the entire request, so rendering can be
performed with full information.