Checkpoint
Moving towards dealing with proper Collection input.  First, a whole lot
of refactoring to get things into the right places.

Changed FieldInfo to FormFieldInfo, and moved the creation of that down
into the Collection.  This requires polluting Collection with knowledge
of Forms, but for now that's probably necessary; later, I'll split out
render/parse classes as decorators or some such.

Also, tweaked the way I'm handling the paragraph marker, so that it
stops giving Eclipse indigestion.