Tweaks to User; Space load problems now tolerated
The main change here is some try/catches added so that trying to load an
inconsistent Space no longer crashes all the way up to the UI.  Instead,
we quietly ignore the problematic Things, and log the Error.  (A bit
dangerous, but necessary from a time perspective.)

Beginnings of implementing Identity correctly, but just the beginning.

Beginnings of email, but again, just the beginning.