Email is now working
Standardized the display of Email Message, including a link to send the
mail.

Removed the Email Cc property, since I realized that I don't know what
it means in the context of mass-customized email.

Introduced the concept of the ?prop= query parameter, which displays the
value of the specified property instead of DisplayText.  This gives us a
much better way to invoke a method.

Added ThingId versions of the lookup functions in SpaceState.