Checkpoint towards Invitation Links
There is now a concept of "Internal Properties" -- if a Property has the
InternalProp meta-property set, it shouldn't be visible in the UI.

Introduced the Identity object.

Added Thing.firstOr(), which lets you specify a default, and ifSet(),
which lets code quickly check a YesNo.

Introduced the IdentityLink method, which Email Messages can use to
embed an invitation to use Querki.  (Does not fully work yet.)

Introduced the ExternalLinkType, AKA URL, which is exactly what it
sounds like -- a type that only accepts valid URLs.  Enhanced the QL
parser so that they work just link internal links in expressions.

Wrote the first-draft Hasher, which will be used at least to hash
invitation links.