Checkpoint towards in-line Editing
We've gotten to the point where the _edit command now injects an
appropriate edit control, although it doesn't actually *do* anything
quite yet.

Enhanced the pipeline so you can pass parameters into calls.  (Although
that is not yet tested.)

Introduced the concept of QLFunction, the abstraction of something that
the pipeline can call.  This is so that we can have partially applied
functions.

Introduced partially applied functions, at least in first draft.

Lifted Methods out to their own source file, since Properties was
getting large.

Massive rework of the Wikitext type, so that we now screen out HTML in
Text fields, but the system can inject HTML that it generates via the
HtmlWikitext class.

Fixed the way collections get rendered.  (Looked like I'd begun to
refactor, but left it off mid-way.)