Added the _section() method
Discovered that I had thought about params wrong -- they *need* to be
pass-by-name in order to work correctly.  (If they are processed before
passing them in, they can't use the context, which misses the whole
point!)

Had to introduce machinery to be able to tell downstream processing to
use the context as a collection, not as individual items.  Did this by
mucking with map() and flatMap(), which I believe is the appropriate way
to handle it, but it means that everybody needs to be good about using
those.

The Context now has a pointer to the Parser, so methods can call back
into it.

Enhanced parameter syntax to cope with whitespace flexibly.