Added Email Message.Recipients
This is the long-desired indirection, so you don't necessarily have to
specify the recipients of an email explicitly, one by one.  This QL
Property returns the list of people to send to, and can be edited as
desired.

Also added PropValue.flatMap(), refactored out of Property.  This
requires explicitly stating the PType, but it's better than nothing.

Fixed a bug in setting up the Edit form -- the DisplayPropVal would
sometimes use a Model instead of the Thing in naming the fields, which
caused things to fail to register when parsing the submitted form.