Very beginnings of access control
For the moment, we just have a horrible hack: anyone who gets in through
an Invite Link can edit the Space.  That'll do for the Wedding RSVPs,
but it needs to be improved ASAP.

Major improvement, though: we are now passing proper User objects
around, not just OIDs.  This lets us be more sophisticated in our
processing of ACLs, and should reduce redundant lookups.