Added the _first method
Also fixed a bug in Wikitext.  It had been that, as soon as you added a
CompositeWikitext into the stream (eg, with _edit), everything following
it would wind up as a CompositeWikitext, screwing up QText rendering.
Fixed that, at the cost of making things even more baroque and messed-up
internally.