We can now define sections by filters
This is the first-draft way to pull apart a list based on its
properties, into different sections.

To do this, added _filter() and _not().

To support that, added YesNoType.toBoolean(), which I'll need for
various predicate work.

Also added Context.firstTyped(), which is kind of an annoying hack, and
flatMapAsContext(), which feels *entirely* wrong but is currently
useful.

Also added _bulleted, since I want that so often.