Recursive functions now work decently
Introduced the _rest function, to compliment _first.

Pulled the concept of partial application up into Property itself, so
that you can use "Thing.Prop" for any property.  In practice, this
mostly matters with QL Properties, which need separate notions of the
defining and received contexts.

Added the notion of context "depth", and capped that at 100, to catch
infinite loops.

We now catch attempts to apply a Property to an empty incoming context,
and fail more gracefully.