Made it not a warning to have a prop in empty context
The wedding invitations were coming up with the error "Trying to apply
property Children in an empty context!".  This was happening for people
with no SO, because we were trying to evaluate [[My SO -> Children]].

This situation is entirely reasonable, though.  I started to make _or
cut processing, but decided that that is incorrect.  Instead, we should
just say that, if you apply a Property to an empty context, the returned
value is empty.  That makes sense, fixes the problem, and allows for
later processing of _if(_isEmpty) for detecting that this has happened
and doing something else.