Replaced the old All Things display
The new display is better in all respects, so removed the old hard-coded
version with this one.  One fewer inconsistency in the system.  You can
now look at the object itself, and it will display the Things in the
invoking Space.  (Sweet!)

Along the way, changed the name of the object to "All Things", since
this is now a front-and-center object, not a buried one.