Rewrite of Wikitext, to cope with HTML blocks
It has been driving me insane that you can't intermix HTML blocks inside
QWikitext, for security reasons.  This was causing problems because
Actuarius was putting in paragraph breaks around all these little bits,
because it didn't understand the full picture of the page.

Completely rewrote the rendering of Wikitext, so that we substitute an
innocuous string for the HTML blocks, and then re-substitute at the end.
This way, those blocks slide past Actuarius.

This will probably need tweaking and enhancement as we go, but it's a
good start.

Also: made the Optional YesNo buttonset appear in-line.  This required
changing the two divs to one span (which was expected), and making a
small CSS hack (which wasn't).  Odds are that it has slightly broken
accessibility, so we'll have to revisit this later.