Tags are beginning to work
Checkpoint, but a big one:

Introduced TagSetType, as a variant of NameType.

Introduced the notion (although nothing's using it yet) that a PType may
imply a Collection.

Added the _showUnknownName property on Spaces, which is predefined on
System.  This is the page to display for the unknown name.

Added the _tagRefs method, which returns all of the Things that have
tags referring to the received context.

Added LinkFromThingBuilder, which is often useful in QList.from().

Introduced rudimentary rendering support for editing Tag Sets.

Major enhancement to Application.withSpace, to allow situation-specific
error handlers that do smarter things than simply pop back to index.  If
thing() gets a "name not found" error, it invokes a variant of the page
to deal.

To support that, finally enhanced the ThingFailed to carry more
information, so it can be properly processed.  Many small mechanical
changes throughout the code to support this properly.