You can now easily see the tags for each Prop
Added the _propsOfType method, which lets you list all of the Tag Set
properties in the Space.  Added that to the default display for a Space.

Made _sort deal with Tag Sets.

Introduced the notion that Properties can declare a special rendering
when you look at them, and more importantly that a *type* can say how to
display properties of that type.  Tag Set properties now default to
listing all of the values of that property.