Link Model is now respected for Tag Set Props
That is, if you define a Tag Set Property with a specific Link Model,
when you go to reify one of those Tags, it will use that Model.

Added PType.matches, to provide a consistent way to compare VT values.
(Needed for NameType, so that it compares case-insensitively.)

Added Property.confirmType, as a way to work around type-math
weaknesses.  Used that to make Space.propsOfType more type-aware.