Fixed _section to process the details properly
The details parameter was still being passed through whole, with the
result that any parameter *other* than Text wasn't being mapped
properly.  Now does context.map() on the details, so that it gets
processed individually over each element.

To make this work the way it used to, had to expose the insertNewlines
parameter through QLParser.contextsToWikitext(), which is probably
appropriate.  In the medium term, it should probably get pushed all the
way up to the public API of _section() as an optional parameter.