Added _isDefined and How It Works
_isDefined is a method that returns true iff the received value is *not*
UndefinedNameType.  Now that we have tags, that becomes necessary a lot
more often.

How It Works is simply a convention, but one I hope to use a lot.  If a
Space provides a How It Works page, that should describe how this Space
functions, so others can learn from it.