Fixed a stack overflow with long unQLText sections
This showed up in the Poker Space, and took a long time to debug: if you
had a long stretch of text (over a screenful) with no QL expressions in
it, the regexes would eventually explode with a stack overflow error.

Splitting it into two distinct regexes seems to fix that, knock on wood.

Along the way, put in a key exception handler, which should deal with a
lot of the errors we've been seeing in-browser.