Sets are now maintained in sorted order
This isn't terribly efficient at this point, but we expect most sets to
be relatively small, so it's not too bad.

As part of this, introduced PType.comp(), so types can declare their
natural sort order.

Note that QSet.from needs to take the PType as a parameter to make this
happen.  Not for the tenth time, I really want ElemValue to contain its
PType.