Introduced better AJAX editing
Added the new setProperty2 entry point, which may entirely replace
setProperty since it is far more powerful.  This is basically a shallow
front end to doEditInternal, so it doesn't cheat like the old
setProperty did.

Changed the thing.html updateValue() handler to use setProperty2.  We
now use jQuery's serialization capability to serialize the values for
the form, and mock up the rest of the form.