Sets of Tags and Links now sort correctly
This was the straw that broke the camel's back, and forced me to begin
pushing Context down from the top.  (Since sorting Links requires
follow(), which requires Context.)  I believe this will become ever-more
common from here on.

Pulled Context and TypedValue out of the QLParser file -- that file had
become *way* too central to way too much code, and it was causing a
stupidly large number of recompiles.