Added edit fallbacks
If the requester can't edit, we shouldn't be displaying edit controls.
So _edit now falls back to simply rendering the property in this case.

Also, added _editOrElse, so you can specify in more detail what you'd
like to show if the requester can't edit this.

Added a little _editOrElse to EmailMessage, so it displays appropriately
to the public.