Checkpoint towards initial security controls
Not fully tested yet, but getting there.

Began fleshing out the AccessControlModule, and made Space fully
dependent on it.

Made Read Thing-specific, as it should be.

Fixed a hideous and horrifying bug that was causing moid()s to be
completely wrong.  We'll need to do some cleanup down the line.

Fixed a bug that prevented you from deleting Set Properties.

Made getLinks recursive -- Links get picked up from Apps.  Eventually,
we should probably have a meta-property that says whether Links and Tags
get searched in the Apps.  For now, we need to be able to find Links, so
that we can have the standard "Public", "Members" and "Owner" concept.