Fixed reifying a Tag (by going to it and editing it)
Previously, when you went to edit a Tag, the initial Display Text was
always the system default.  Changed to use the Space's current Show
Unknown Name text, which is actually the right value.

Also fixed Show Unknown Name to be a Large Text instead of a QLType --
not sure why I had it that way, but it was clearly incorrect.