Got _modTime working
This is an overlarge checkin, due once again to GitHub for Windows
crashing and burning; *sigh*.

We now load in the new "modified" column from the DB when loading the
Space.  All Things have a modTime -- a real one if possible, synthesized
if not.

Introduced the new TimeModule, which is where all time-related
functionality will live.  This uses Joda-time (wrapped by nscala-time)
as our official internal Time format.

As part of that, added the _modTime method, which produces the modTime
for the received Thing.

Enhanced _sort to take an optional parameter, which is evaluated for
each received Thing, and establishes what to sort on.  This means that
we now actually use pType.comp() when we should do so.

Tweaked LinkType.followLink() to be more robust, having noticed that its
current form very susceptible to exceptions.