Added the _desc method
This proved to be surprisingly easy, using the machinery I already had.
_desc turns out to be a Type modifier, which returns the received value
with one tweak: the Type is changed to have a reversed sort order.

Fixed a bug in _sort, which was breaking the Context chain.

Enhanced the error-catching in QL processing, so that we do print the
call stack if we have one.