Added a bunch of methods
Moving towards being able to generate documentation.

Fixed _allProps so that it actually displays the properties in the
received Space; previously, it was always displaying the props of the
*context* Space.

Added the _oid method.

Added the _kind method.

Added the _currentSpace method, which returns the *root* context's
Space.  (As opposed to _space, which returns the Space of the received
Thing.)  This is needed to be able to list the Properties and Things of
System without needing to be *in* System.