Fixed the handling of underscores in link names
Actuarius wasn't handling [_myThing](_myThing) correctly -- it was
treating the underscores as italics, and failing to recognize it as a
link.  Changed the code to allow all legal URL characters in links, and
to favor that instead of Markdown'izing the text.  Added a new test line
to confirm the fix.