Tweaked _section to not demand individual-element
Realized that _section's semantics (which are very old) are by now
obsolete.  Whether a parameter is interpreting the input as collection
or elements is a function of the *phrase*, not the caller.  In fact,
_section was double-mapping in the usual case, since the Text processor
was already mapping.

This change means that you can now put list-oriented methods into
DETAILS (eg, _commas), and it Just Works.  Yay!

Also, enhanced QLRequestContext to be more forgiving of actually being
used.  It had been causing View Source to crash, as a side-effect of
pathAdjustments().