Checkpointing a bunch of half-finished stuff
GfW crapped out mid-thought, so catching stuff up here.

Introduced the beginnings of a sidebar for Sharing and Security, since I
expect that page to get complex.

Added Hasher.sign(), but that's not tested yet.

Refactored makeWikiLink, to begin to unbreak those paths preparatory to
invitations.

Began to fix a horrible bug I noticed the other day, in Property
application.