Fixed the bug about displaying complex Texts
Basically, the problem had been that if I have MyThing, and define
Summary on it to be a complex summary using MyThing's other Properties,
it didn't work as expected.  If Summary looked like this:

[[""... the actual contents...""]]

it would work, but if I simply said:

... the actual contents...

only the first Thing would display.

The bug was subtle, but boiled down to the fact that
applyToIncomingThing() only works on the *first* Thing in the incoming
Context.  Figuring out where to inject a map proved hairy, but the
safest and most correct answer appears to be treating QLText almost
exactly the same way we already treat QL, which does make a certain
amount of sense...