Added displayOpt to the render signature.
Checkpoint: still a bunch more to so.

The key notion here is that, instead of handling ____ as an ad-hoc
concept in QLParser, we're going to build it all the way down into
PType, so that any Type can potentially have a "display parameter".  For
example, this will be lovely for DateTime, as a way to specify the
rendering param.

Also, renamed render() to wikify(), because adding the default parameter
turns out to conflict with Thing.render() -- the compiler isn't clever
enough to tell that the other parameters distinguish them sufficiently.