Getting to the point
Finally done with that refactor.  The rendering of ____ and __stuff__ is
now handled entirely consistently, with the Type responsible for
deciding how it wants to deal with the optional display parameter.

This does fix one small bug: previously, if you linked to an unknown
name with display text, it did not show up correctly highlighted, due to
duplicate code paths.  It now does.