Only the Owner can change Permissions
Decided to play this one "fair", more or less.  Introduced much of the
powertrain for what amounts to column-level security: when you go to
change a Thing, it checks each changed Property, and asks whether you
are allowed to modify that Property.

The only cheat is that, for the moment, we hardcode it so that if the
Property is a Permission, only the Owner can change it; otherwise,
anyone can do so.  But we should be able to get to full column-level
security simply by adding more machinery in the AccessControlModule.