Bugfix: can now join private Spaces
It turned out that inviting someone to join a non-Public Space *totally*
failed, in many ways.  The core issue is that much of the code depended
on fetching the SpaceState, which of course failed if you can't read
that State.

Wound up doing major surgery to the RC code.  Changed the way that
login-to-join works, because I just couldn't find a clean way to do it.
Introduced the notion that the RC carries the raw owner/space request
info, so it can be used *before* fetching the Space.

Disabled the old IdentityLoginChecker.  This means that the invites to
the old Wedding Site are now broken.  Sad, but unsurprising.