Added Pick Lists
New feature, so that I can build the Shopping List Space for Kate.

_editAsPickList is a new command that you can use on Tag/LinkSets that
have a Link Model.  It renders them instead as a list of checkboxes of
*all* the candidates, so you can easily add the things you want to the
list.

To implement this, added the concept of "specializations" for input
rendering.  This is clunky, and the factoring needs thought, but should
do for now.

Put the new feature into a new Rendering Module, so as not to clutter up
the global ID space any worse.