Tweaks to withAdd
We now add the new items by *display name*, not name, since that is
usually what we want.

You can now just hit Return to add the new item, instead of pressing the
button.  By and large, you're going to want the button on a mobile
device, or Return on a desktop.