Fixed pick lists in IE9
It turns out that you are not allowed to put a <form> inside a <p> in
HTML, and IE9 deals with this in a particularly heinous way: it
restructures the HTML in a way that breaks my forms.

Solution is to stop using <p> altogether, replacing it with <div
class="para">.  This feels a bit dramatic, but I'm not coming up with
any actual problems with it, and it fixes the problem.