Got TagThing pages working again
A subtle bug had arisen.  Unsurprisingly, this was in an old cheat:
LinkType.followLink was calling firstTyped, which throws if the type is
wrong.  (Which in this case it was.)  Fixed by the simple expedient of
switching that to firstAs, the more modern and correct version.

The implication (since this was being called under pathAdjustments) is
probably that you can't put subpaths under a node that doesn't exist.
We can live with that for now.

Wound up beefing up QLParser.logContext to track this down, and putting
it on a config variable as it always should have been.  Also, made it
work in more circumstances -- the return value can be anything
DebugRenderable.  There is a general mechanism trying to break out here.