Fixed the "failed to add to a list" problem
This was a long-standing design bug that's been driving me nuts.  If a
List was inherited, it is of course initially supposed to be disabled
until you press the Edit button.  But the Add button was *not* disabled,
so you could go adding things, which would get picked up because you
weren't allowed to edit it!

Introduced a "delegating" mechanism, so that disabling a List now
actually disables the Add button.  Not perfect, but it'll do for now.