Refactored RC -> Owner fields
These have been going through runtime lookups in User every time we
access ownerHandle and ownerName.  That's insane for several reasons: it
binds RC to a database operation, it's inefficient, and it's just plain
incorrect for a case class.

Changed things so that ownerIdentity is now bound to the SpaceState, and
can be looked up directly from there; handle and name fall out of that.
This allows us to clean up a lot of code.