Big refactor of RequestContext
In preparation for upcoming unit testing, finally dealt with a
long-standing smell: the fact that RequestContext (which is passed all
over, hither and yon) contains Play information.  This was causing
controllers to bleed into the backend code, which is strictly Bad.

So I reworked that collection of classes, slimming RequestContext itself
down to just the basics, and deriving a PlayRequestContext subclass with
all of the Play-level stuff.  This confines the HTTP details to the
front end, where they belong, and removes many grody references to
controllers from the back end.

The new, smaller RequestContext now lives in querki.values, where it
belongs.  It might yet wind up merged with QLContext, but we'll see.

Along the way, moved NavSection up to controllers, which is really where
it always belonged.

Finally deleted the old IdentityLoginChecker, which was now too much in
the way, and which was disabled a while ago.