Lifted Space to new querki.spaces package
Checkpoint -- beginning to refactor Spaces, one bit at a time.

Created querki.spaces, and moved Space there, since that is the
least-dangerous piece of Spaces.

Discovered that oidMap was living in there, which is silly.  Moved it to
a first-class constructor of its own, housed near OID, which makes more
sense.  This actually accounts for most of the changes.