Refactorings of CreateSpace
Checkpoint.  Hasn't moved yet, but lots of restructuring so that it can.

Introduced ThingError finally, so I can begin to move towards
better-managed error messages from the back end.

Introduced UnexpectedPublicException, to cover non-public exceptions.

Fixed a bug that would show a RSOD after Space creation.  (Because of an
obsolete attempt to create the Space Actor.)