Beginning of Terms of Service
Checkpoint.  Introduced the TOSModule, which doesn't do anything but
does have a place in the page-handling pipeline.  Added the tosVersion
column to the User table and class.

Removed the SpaceSpecificUser, since there is no longer any reason to
maintain it.  Began to shift _me to make sense with the new world.