Test refactorings
Changed TestThing so that it auto-registers, which makes it a bit easier
to use, although slightly more dangerous.  (Need to use strict vals now,
so it is easier to accidentally get reference loops in creation -- need
to use explicit OIDs more often, mostly.)

Changed the test setup to make it possible to introduce little,
test-specific Spaces that introduce their own Things.  Not as concise as
I'd like, but not terrible, and allows me to keep the factoring under
control.