Made it easier to embed pictures in pages
Added the ability to just say [[My Photo]], and have it render in place,
as you would expect it to do.  Basically, a Link to an Attachment
renders as a photo.  (This will need enhancement when we add more
Attachment types.)

Properly unit-tested.  Did a little more test refactoring to support
this.