Fixed the display of pictures in All Things
My earlier checkin had broken All Things: we were now displaying the
picture itself there, instead of the link to it.

Fixed things by being more consistent with the rest of the world --
Attachment is now a subclass of ThingState, and overrides qlApply to
only do the auto-insertion thing in this case.  I suspect this still
isn't correct, and that we should have Photo as a *PType*, but it's
better...